backend/requirements.txt
fastapi[standard]
uvicorn[standard]
alembic

pydantic
pydantic-settings
python-dotenv

sqlalchemy[asyncio]
asyncpg

google-adk
litellm

e2b
e2b-code-interpreter
